do not alter header string unless it is set
Given the message:

  Foo:   bar
   baz
     quux
  Bar: 1

  Body

This program:

  Email::Simple->new($message)->as_string

Used to produce:

  Foo: bar baz quux
  Bar: 1

  Body

This is all well and good as far as lexical content goes, but there are
reasons to avoid altering the header representation.  For example, if a
DKIM signature is present, using "simple" canonicalization, the headers
must not be altered in any way, including ways that are, one would
think, entirely irrelevant to the content's meaning.

With this change, ->header continues to return the same thing, but
->as_string will use the existing form.  When a header is set, it only
gets a "parsed" value, and will re-fold every time, etc.