convert vertical whitespace in headers into horizontal
You should never do this:

  Email::Simple->create(
    header => [ one => "two\nthree: four" ],
    ...
  );

It's not good.  What used to happen is that the email message ended
up with both a "one" header and a "two" header, because the \n was
left as-is.  This is obviously no good.

Further, don't do this:

  Email::Simple->create(
    header => [ one => "two\n\nthree: four" ],
    ...
  );

Now the body starts with "three: four" followed by maybe some
more headers.

From here on out, vertical whitespace in the header causes a
warning and becomes horizontal whitespace.  Any sequence of
vertical whitespace in a row becomes one space (U+0020).

In the future, this might change: different compaction, or
fatalization.