Fix iterating arrayref $headers when it is empty
Expression (0 .. int($#$headers / 2)) expand to list (0) when $headers is [].

It caused that both methods headers() and header_names() in list context
returned (undef) when there was no header defined. And it throw warning:
Use of uninitialized value $_ in lc at Email/Simple/Header.pm line 138.

Expression (0 .. @$headers / 2 - 1) expand to list () when $headers is [].

So it should fix it. Warnings were thrown also for header_raw() and
header_raw_set() methods.